Fix: AndPermission can't return self for & operator
Returning self in & is not intended usage for the operator.
The operator is meant to be used as:

```
perm = Perm1() & Perm2() & Perm3()
```

However, if we wanted to reuse permissions above we may want to use:

```
perm_other = perm & Perm3()
```

Before this commit such an operation would not only alter `perm_other`
but also `perm`. Moreover, `perm_other` would be identical to `perm`,
meaning `perm is perm_other`.